requirements.txt
absl-py==2.0.0
addict==2.4.0
ansi2html==1.9.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.0
fonttools==4.47.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.26.3
oauthlib==3.2.2
open3d==0.18.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0
git+https://github.com/graphdeco-inria/diff-gaussian-rasterization.git
